10/09/2023/00:22

+++ test.sql
@@ -1,7 +1,7 @@
 -- 9 new
 CREATE TABLE test (
     test_id Int NOT NULL AUTO_INCREMENT ,
-    score Int ,
+    score FLOAT ,
     grade Varchar (2) ,
     student_id Varchar (13) ,
     PRIMARY KEY (test_id) ,
